requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.18.2
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
serializers==0.2.4
sqlparse==0.4.2
whitenoise==6.2.0
